backend/app/cli.py
import click
from flask import Flask

from app.extensions import db


def register_cli(app: Flask) -> None:
    @app.cli.command("init-db")
    def init_db():
        db.create_all()
        click.echo("Database tables ensured.")

