backend/requirements.txt
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.1
Flask-JWT-Extended==4.6.0
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.32
marshmallow==3.21.3
python-dotenv==1.0.1
psycopg[binary]==3.2.1
redis==5.0.8
celery==5.4.0
gunicorn==22.0.0
pytest==8.3.2

